Add scrape-prep
This replaces part of the prior-removed lambda entrypoint in commit
e9843d4b with a basic script which initiates a scrape. This one is
very simple and only gets the vendor data from dynamodb, then
passes said data into SNS to trigger the lambda which performs a
scrape with requests.